// src/contracts/newport/typechain/OptionMarketWrapper.ts
export declare namespace OptionMarketWrapperWithSwaps {
  export type OptionMarketContractsStruct = {
    quoteAsset: PromiseOrValue<string>;
    baseAsset: PromiseOrValue<string>;
    optionToken: PromiseOrValue<string>;
    liquidityPool: PromiseOrValue<string>;
    liquidityToken: PromiseOrValue<string>;
  };

  export type OptionMarketContractsStructOutput = [
    string,
    string,
    string,
    string,
    string
  ] & {
    quoteAsset: string;
    baseAsset: string;
    optionToken: string;
    liquidityPool: string;
    liquidityToken: string;
  };

  export type OptionPositionParamsStruct = {
    optionMarket: PromiseOrValue<string>;
    strikeId: PromiseOrValue<BigNumberish>;
    positionId: PromiseOrValue<BigNumberish>;
    iterations: PromiseOrValue<BigNumberish>;
    setCollateralTo: PromiseOrValue<BigNumberish>;
    currentCollateral: PromiseOrValue<BigNumberish>;
    optionType: PromiseOrValue<BigNumberish>;
    amount: PromiseOrValue<BigNumberish>;
    minCost: PromiseOrValue<BigNumberish>;
    maxCost: PromiseOrValue<BigNumberish>;
    inputAmount: PromiseOrValue<BigNumberish>;
    inputAsset: PromiseOrValue<string>;
  };

  export type OptionPositionParamsStructOutput = [
    string,
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber,
    number,
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber,
    string
  ] & {
    optionMarket: string;
    strikeId: BigNumber;
    positionId: BigNumber;
    iterations: BigNumber;
    setCollateralTo: BigNumber;
    currentCollateral: BigNumber;
    optionType: number;
    amount: BigNumber;
    minCost: BigNumber;
    maxCost: BigNumber;
    inputAmount: BigNumber;
    inputAsset: string;
  };

  export type ReturnDetailsStruct = {
    market: PromiseOrValue<string>;
    positionId: PromiseOrValue<BigNumberish>;
    owner: PromiseOrValue<string>;
    amount: PromiseOrValue<BigNumberish>;
    totalCost: PromiseOrValue<BigNumberish>;
    totalFee: PromiseOrValue<BigNumberish>;
    swapFee: PromiseOrValue<BigNumberish>;
    token: PromiseOrValue<string>;
  };

  export type ReturnDetailsStructOutput = [
    string,
    BigNumber,
    string,
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber,
    string
  ] & {
    market: string;
    positionId: BigNumber;
    owner: string;
    amount: BigNumber;
    totalCost: BigNumber;
    totalFee: BigNumber;
    swapFee: BigNumber;
    token: string;
  };

  export type StableAssetViewStruct = {
    id: PromiseOrValue<BigNumberish>;
    token: PromiseOrValue<string>;
    decimals: PromiseOrValue<BigNumberish>;
    symbol: PromiseOrValue<string>;
    balance: PromiseOrValue<BigNumberish>;
    allowance: PromiseOrValue<BigNumberish>;
  };

  export type StableAssetViewStructOutput = [
    number,
    string,
    number,
    string,
    BigNumber,
    BigNumber
  ] & {
    id: number;
    token: string;
    decimals: number;
    symbol: string;
    balance: BigNumber;
    allowance: BigNumber;
  };

  export type MarketAssetViewStruct = {
    id: PromiseOrValue<BigNumberish>;
    market: PromiseOrValue<string>;
    token: PromiseOrValue<string>;
    decimals: PromiseOrValue<BigNumberish>;
    symbol: PromiseOrValue<string>;
    balance: PromiseOrValue<BigNumberish>;
    allowance: PromiseOrValue<BigNumberish>;
    isApprovedForAll: PromiseOrValue<boolean>;
  };

  export type MarketAssetViewStructOutput = [
    number,
    string,
    string,
    number,
    string,
    BigNumber,
    BigNumber,
    boolean
  ] & {
    id: number;
    market: string;
    token: string;
    decimals: number;
    symbol: string;
    balance: BigNumber;
    allowance: BigNumber;
    isApprovedForAll: boolean;
  };

  export type LiquidityBalanceAndAllowanceStruct = {
    token: PromiseOrValue<string>;
    balance: PromiseOrValue<BigNumberish>;
    allowance: PromiseOrValue<BigNumberish>;
  };

  export type LiquidityBalanceAndAllowanceStructOutput = [
    string,
    BigNumber,
    BigNumber
  ] & { token: string; balance: BigNumber; allowance: BigNumber };
}